Implements fmt::LowerHex and fmt::UpperHex for GenericArray<u8, N>
Since this crate is very useful for hash digest libraries it's also very
useful to output value of the array in hex.

This implementation allocates a buffer for hex data on the stack and
uses `write_str`. This method is about 2x faster comparing to writing
every char to the buffer using `write_char`.

It's also 20-30 percent faster than using `rustc_serialize::hex::ToHex`.
Note, the latter also does a memory allocation which we avoid in this
implementation.

For arrays larger than 1024 bytes we don't allocate the whole hex data
on the stack but process data in 1024 bytes chunk (with 2048 bytes
buffer). The size of the chunk is chosed arbitrarily but it's smaller
than standard buffer size used for `BufReader` and `io::copy`.