Add optional support for Zeroize
zeroize is the most popular crate for zeroizing memory after use.
generic-array is the most popular crate for representing bytes in
cryptographic implementations which can't use an allocator.

Issues have been opened in RustCrypto about zeroizing instances of
generic-array, for instance here:
https://github.com/RustCrypto/hashes/issues/87

However, sometimes you want to return generic-array from an API
while also commiting it to be zeroized after use, because the caller
might forget to do this on some code path. The natural way to do that
in zeroize crate is the `Zeroizing` wrapper. However, `Zeroizing`
cannot be used with `generic-array` unless `generic-array` implements
the `Zeroize` trait.

The easiest way to do that is create an optional dependency on
`Zeroize` and put the implementation in a conditionally-compiled
module, as we did in this commit.