Fully initialize new iterator upon clone (crudely)
We need something to put into the contents of the new GenericArray to
back the iterator upon clone. The perfect solution would be to change
the element type to MaybeUninit, but that invades the trait bounds too
much. This solution is somewhat disconcerting, but unlike the previous
implementation's use of assume_init(), I cannot get miri to reject it.