Trait for default const values
Utilises the `const-default` crate (behind a feature gate) to introduce
a trait defining an associated constant of the `Self` type.  An
implementation is provided for `GenericArray`, providing means of
constructing default `GenericArray`s in const context.

This plugs a gap that exists even after #130 was merged, as generic type
parameters cannot currently be referenced in const context (e.g. if the
repetition length provided to the `arr!` macro comes from a type
parameter).